refactor(config): keep the config core free of sinks (refs #2426)
CI's `Dependency boundaries` job went red on the first push: wiring the
loaders onto the core added NINE import cycles. Every config loader sits
downstream of `file-utils.ts`, so importing the core closed a loop back
through `config-warn -> degradation-ledger -> extension-log -> file-utils`.

Six of the nine were the core's own doing rather than the wiring's, and
are removed rather than blessed:

- `resolveConfig` moves out of `config-core/index.ts` into
  `config-core/resolve.ts`. The barrel re-exports the whole vocabulary,
  so importing the front door also pulled `process-spec.js` and, with
  it, `project-trust.js` — three cycles for a dependency a resolver
  never uses. `index.ts` re-exports it; the public surface is unchanged.
- `reportMigrationRecords` moves to `clients/config-resolve.ts` as
  `reportPiLensConfigRecords`, and `DEGRADATION_ENTRIES_PER_KIND` moves
  to the `ledger-bounds.js` leaf. `config-core` documents itself as
  "Pure: no state, no I/O, no ledger writes" while importing the warn
  seam and the ledger; the import graph disagreed with the comment.
  `ledger-bounds.ts`'s own doc already gives the reason — a producer
  that needs the BOUND must not reach through the ledger, because tests
  routinely mock it wholesale, and three suites broke exactly that way
  on the first push.

Nothing under `config-core/` imports a sink now.

The remaining three are the pre-existing cycle class #2418 already
blessed, reshaped: `config-resolve.js` now sits where each loader sat
directly. Baseline 31 -> 33 (2 reshaped, 2 net-new config-warn edges),
with the reasoning recorded in the known-violations file beside #2418's
and #2173's.

Refs #2426
