+++ tests/config/dependency-boundaries.test.ts
@@ -95,6 +95,19 @@ describe("dependency boundary governance", () => {
 		// chain, so that edge closes a baseline-class cycle rather than
 		// introducing new coupling — the same reasoning entry 30 was accepted
 		// under (#2173). Reduction stays tracked in #2125.
+		//
+		// #2426: 31 -> 33. Two entries RESHAPED (the loaders resolve through the
+		// shared `clients/config-resolve.js`, so it sits on the pre-existing
+		// config-warn -> degradation-ledger -> extension-log -> file-utils cycle
+		// where `project-lens-config.js` sat directly) and two net-new
+		// config-warn edges, both the shared resolver reporting through the SAME
+		// seam each loader already used. The same PR REMOVED six cycles it would
+		// otherwise have added: `resolveConfig` moved out of the `config-core`
+		// barrel so a resolver no longer drags `process-spec` -> `project-trust`,
+		// and `config-core` lost both of its sink imports
+		// (`reportMigrationRecords` -> `config-resolve.js`,
+		// `DEGRADATION_ENTRIES_PER_KIND` -> the `ledger-bounds.js` leaf), which is
+		// what that module's own doc always claimed of it.
 		const baseline = parseJson5(
 			readFileSync(
 				resolve(repoRoot, ".dependency-cruiser-known-violations.json"),
@@ -102,6 +115,6 @@ describe("dependency boundary governance", () => {
 			),
 		) as unknown[];
 
-		expect(baseline).toHaveLength(31);
+		expect(baseline).toHaveLength(33);
 	}, 30_000);
 });
